+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.5.1 2026-04-04
 ### Breaking change
 - Reminder: Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Configuration#how-to-delete-and-re-add-the-solarcharger-integration) if upgrading from v0.4.3 or prior.
+### Fix
+- Fixed issue with SolarCharger not stopping at sunset when in paused state.
 ### Improvement
 - Improved semaphore code to wake up and update HA on device presence detection.